Created Sensor List. Updates to Sensor
Comments required on SensorList. Sensor now returns unsigned long as opposed to longs as sensor values are not negative and int is smaller.

+++ Sensor.cpp
@@ -31,7 +31,7 @@ Sensor::~Sensor() {
 
 }
 
-long Sensor::updateDataViaPin(int samples) {
+unsigned int Sensor::updateDataViaPin(int samples) {
     digitalWrite(powerPin, HIGH);
     delay(10);
 	this->data = 0;
